refactor: pull chunkedOrphanDelete into shared module + defensive fixes
Audit follow-ups across the PR. Three changes:

1. **Extract orphan cleanup to src/db/orphan-cleanup.js.** Both
   src/db/scanner.mjs (end-of-scan cleanup, child process) and
   src/util/admin.js removeDirectory (post-vpath-delete cleanup, main
   process) had near-identical copies of chunkedOrphanDelete + the
   three orphan-id queries. Now both call cleanupOrphans(db) — one
   import line at each site, the chunk-size constant + SQL strings +
   loop semantics live in one place. Side effect: util/admin.js's
   removeDirectory now also cleans orphan genres (the scanner already
   did; admin had been silently skipping the third table).

   The Rust scanner keeps its own copy in rust-parser/src/main.rs —
   different language, separate process, importing across that
   boundary isn't an option. Comments stay in lockstep.

2. **triggerForLibrary: per-destination try/catch.** A throw from
   enqueue() (today only theoretical — it catches its own DB errors)
   would have killed the loop and prevented every later destination
   from being triggered. Cheap defensive isolation.

3. **bufferLines: 'error' handler on the stream.** Child-process
   pipes can emit 'error' (broken pipe, EIO from a force-killed
   child). Without a listener Node treats it as unhandled and tears
   the parent process down. Logs at warn and lets the existing 'close'
   handler do cleanup as normal.

Also considered and intentionally NOT changed:
  - getAdminStats().vpaths is now always 0-or-1 entries; could be
    flattened to a string. Backwards-compat with hypothetical external
    consumers wins — the array shape is harmless.
  - JS scanner emits scanComplete BEFORE cleanup; Rust emits AFTER.
    Both reach the parent-side close handler the same way; the order
    is cosmetic. Leaving each as-is preserves the JS path's "report
    success even if cleanup fails" failure-mode.
  - recordFileError in worker.mjs writes one stderr line per failure.
    A pathologically-broken backup could spam thousands. Sample-cap
    would help but loses information; deferring until someone
    actually hits it.
